requirements.txt
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
Created on Tue Mar 24 16:08:16 2026

@author: elnarbaynazarov
"""

Flask
pillow
numpy
opencv-python
pytesseract
